添加注释、修改文件结构
