+++ sports meet/qt/test.pro
