+++ sports meet/qt/widget.cpp
