+++ sports meet/qt/widget.h
