--- 运动会问题/qt/myqtextedit.h
@@ -1,17 +1,0 @@
-#ifndef MYQTEXTEDIT_H
-#define MYQTEXTEDIT_H
-
-#include<QLineEdit>
-#include<QPlainTextEdit>
-class myQTextEdit : public QPlainTextEdit
-{
-    Q_OBJECT
-public:
-    void setedit( QLineEdit *edit );
-public slots:
-    void mysetvalue();
-private:
-    QLineEdit *edit;
-};
-
-#endif // MYQTEXTEDIT_H
